Refactored the ui components to match one naming style
