+++ captura/ui/TemplateDelegate.py
